+++ lib/src/main/java/hr/avrbanac/preequalization/lib/struct/DefaultPreEqData.java
@@ -19,7 +19,7 @@ import java.util.List;
  * </h3>
  *
  * <p>
- * Each tap is coupled with delay element (for docsis 2.0 and 3.0 - symbol period T; and for docsis 1.1 it was T/2 and T/4).
+ * Each tap is coupled with delay element (for DOCSIS 2.0 and 3.0 - symbol period T; and for DOCSIS 1.1 it was T/2 and T/4).
  * In the ranging process the CM sends a ranging request message (RNG-REQ) to the CMTS, which uses a known portion of this message
  * (preamble) to determine the quality of the received signal and also the adjustments needed to better compensate the upstream distortion.
  * Thus, the response message is sent (RNG-RSP), containing a set of 24 coefficients and a parameter that defines whether it is a set
@@ -29,19 +29,19 @@ import java.util.List;
  *
  * <p>
  * RNG-RSP message is a TLV encoded message with following information: timing, frequency, power level, equalization adjustment, set
- * or adjust parameter and ranging status. For docsis 2.0 and 3.0 equalization coefficients are identified as type 09 (docsis 1.1 has
+ * or adjust parameter and ranging status. For DOCSIS 2.0 and 3.0 equalization coefficients are identified as type 09 (DOCSIS 1.1 has
  * type 04). RNG-RSP and RNG-REQ messages are linked by the SID number (service ID) - present in both messages. Additionally, channel ID is
  * also provided.
  * </p>
  *
  * <p>
  * RNG-REQ message is a TLV encoded message with known portion of the message (preamble) which is used to determine needed adjustments.
- * Roughly, RNG-REQ message is divided into 3 parts: ramp-up with preamble, docsis payload and FEC info/parity woth ramp-down. Docsis
+ * Roughly, RNG-REQ message is divided into 3 parts: ramp-up with preamble, DOCSIS payload and FEC info/parity woth ramp-down. Docsis
  * payload part carries SID, channel ID and pending till complete info (used to inform CMTS that ranging adjustment is not yet completed).
  * </p>
  *
  * <p>
- * Historically, docsis 1.0 and 1.1 had some problems with pre-eq but that is beyond the scope of this project.
+ * Historically, DOCSIS 1.0 and 1.1 had some problems with pre-eq but that is beyond the scope of this project.
  * </p>
  * <hr/>
  * <h3>
@@ -50,7 +50,7 @@ import java.util.List;
  *
  * <p>
  * The maximum delay compensation that can be achieved using pre-eq is equal to the delay between adaptive equalizer's main tap and last
- * adaptive equalizer tap. For docsis 2.0 and 3.0, the delay or spacing between each adaptive equalizer tap location is equal to the symbol
+ * adaptive equalizer tap. For DOCSIS 2.0 and 3.0, the delay or spacing between each adaptive equalizer tap location is equal to the symbol
  * period. Typical implementations have main eq. tap in eight. position out of 24-tap delay line. This gives us max delay of 16T.
  * </p>
  *
@@ -133,9 +133,35 @@ public class DefaultPreEqData implements PreEqData {
     private static final int TAP_COUNT = 24;
     private static final int COEFFICIENT_PER_SYMBOL = 1;
 
+    /**
+     * Normalized pre-eq input string.
+     */
     private final String preEqString;
+
+    /**
+     * Main tap index fetched from pre-eq header data. THIS IS AN ACTUAL INDEX, NOT AN ARRAY INDEX (use 1 less for array).
+     */
     private final int mainTapIndex;
+
+    /**
+     * Parsed complex coefficients corresponding to 24 energy taps.
+     */
     private final List<Coefficient> coefficients = new ArrayList<>();
+
+    // The following are the key metrics:
+    private final long lMTE;
+    private final long lMTNA;
+    private final long lMTNE;
+    private final long lPreMTE;
+    private final long lPostMTE;
+    private final long lTTE;
+    private final double dMTC;
+    private final double dMTR;
+    private final double dNMTER;
+    private final double dPreMTTER;
+    private final double dPostMTTER;
+    private final double dPPESR;
+    private final double dPPTSR;
 
     public DefaultPreEqData(final String rawInputPreEqString) {
         this.preEqString = rawInputPreEqString
@@ -159,37 +185,178 @@ public class DefaultPreEqData implements PreEqData {
         }
 
         for (int i = 4; i < bytes.length; i += 4) {
-            coefficients.add(new DefaultCoefficient(Arrays.copyOfRange(bytes, i, i + 4)));
+            coefficients.add(new DefaultCoefficient(Arrays.copyOfRange(bytes, i, i + 4), i / 4));
         }
+
+        lMTE = coefficients.get(mainTapIndex - 1).getEnergy();
+        lPreMTE = calculateEnergyForTaps(1, mainTapIndex - 1);
+        lPostMTE = calculateEnergyForTaps(mainTapIndex + 1, TAP_COUNT);
+        lTTE = lPreMTE + lMTE + lPostMTE;
+        dMTC = 10 * Math.log10(1d * lTTE / lMTE);
+        dMTR = 10 * Math.log10(1d * lMTE / (lPreMTE + lPostMTE));
+        dNMTER = 10 * Math.log10(1d * (lPreMTE + lPostMTE) / lTTE);
+        dPreMTTER = 10 * Math.log10(1d * lPreMTE / lTTE);
+        dPostMTTER = 10 * Math.log10(1d * lPostMTE / lTTE);
+        dPPESR = 10 * Math.log10(1d * lPreMTE / lPostMTE);
+        dPPTSR = 10 * Math.log10(1d * coefficients.get(mainTapIndex - 2).getEnergy() / coefficients.get(mainTapIndex).getEnergy());
+
+        lMTNA = Math.round(Math.pow(2, Math.ceil(Math.log(Math.sqrt(lTTE)) / Math.log(2))) - 1);
+        lMTNE = lMTNA * lMTNA;
     }
 
+    public long calculateEnergyForTaps(
+            final int startTap,
+            final int endTap) {
+
+        long energy = 0L;
+        for (int i = startTap - 1; i < endTap; i++) {
+            energy += coefficients.get(i).getEnergy();
+        }
+
+        return energy;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public String getPreEqString() {
         return preEqString;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public List<Coefficient> getCoefficients() {
         return coefficients;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public int getTapCount() {
         return TAP_COUNT;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public int getMainTapIndex() {
         return mainTapIndex;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public int getCoefficientPerSymbol() {
         return COEFFICIENT_PER_SYMBOL;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public long getMTE() {
-        return coefficients.get(mainTapIndex - 1).getEnergy();
+        return lMTE;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public long getMTNA() {
+        return lMTNA;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public long getMTNE() {
+        return lMTNE;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public long getPreMTE() {
+        return lPreMTE;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public long getPostMTE() {
+        return lPostMTE;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public long getTTE() {
+        return lTTE;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public double getMTC() {
+        return dMTC;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public double getMTR() {
+        return dMTR;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public double getNMTER() {
+        return dNMTER;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public double getPreMTTER() {
+        return dPreMTTER;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public double getPostMTTER() {
+        return dPostMTTER;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public double getPPESR() {
+        return dPPESR;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public double getPPTSR() {
+        return dPPTSR;
     }
 }