+++ CHANGELOG.md
@@ -17,6 +17,7 @@ At the moment, both library and application are under development.
 - lib log configuration (resources)
 - pre-eq exception class and struct package with structure interfaces
 - default implementation classes and main parsing logic
+- key metrics calculation
 
 ### Deprecated
 - nothing
@@ -29,6 +30,7 @@ At the moment, both library and application are under development.
 
 ### Tested
 - default coefficient creation
+- default pre-eq data creation with structure output
 
 ### Security
 - nothing